abstract: Pools/Config
