+++ src/contexts/Pools/Config.tsx
@@ -1,0 +1,129 @@
+// Copyright 2022 @paritytech/polkadot-staking-dashboard authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+import BN from 'bn.js';
+import React, { useState, useEffect } from 'react';
+import { APIContextInterface } from 'types/api';
+import { PoolsConfigContextState } from 'types/pools';
+import * as defaults from './defaults';
+import { useApi } from '../Api';
+import { rmCommas } from '../../Utils';
+
+export const PoolsConfigContext =
+  React.createContext<PoolsConfigContextState | null>(null);
+
+export const usePoolsConfig = () => React.useContext(PoolsConfigContext);
+
+export const PoolsConfigProvider = ({
+  children,
+}: {
+  children: React.ReactNode;
+}) => {
+  const { api, network, isReady } = useApi() as APIContextInterface;
+  const { features } = network;
+
+  // whether pools are enabled
+  const [enabled, setEnabled] = useState(0);
+
+  // store pool metadata
+  const [poolsConfig, setPoolsConfig]: any = useState({
+    stats: defaults.stats,
+    unsub: null,
+  });
+
+  // disable pools if network does not support them
+  useEffect(() => {
+    if (features.pools) {
+      setEnabled(1);
+    } else {
+      setEnabled(0);
+      unsubscribe();
+    }
+  }, [network]);
+
+  useEffect(() => {
+    if (isReady && enabled) {
+      subscribeToPoolConfig();
+    }
+    return () => {
+      unsubscribe();
+    };
+  }, [network, isReady, enabled]);
+
+  const unsubscribe = () => {
+    if (poolsConfig.unsub !== null) {
+      poolsConfig.unsub();
+    }
+  };
+
+  // subscribe to pool chain state
+  const subscribeToPoolConfig = async () => {
+    if (!api) return;
+
+    const unsub = await api.queryMulti(
+      [
+        api.query.nominationPools.counterForPoolMembers,
+        api.query.nominationPools.counterForBondedPools,
+        api.query.nominationPools.counterForRewardPools,
+        api.query.nominationPools.maxPoolMembers,
+        api.query.nominationPools.maxPoolMembersPerPool,
+        api.query.nominationPools.maxPools,
+        api.query.nominationPools.minCreateBond,
+        api.query.nominationPools.minJoinBond,
+      ],
+      ([
+        _counterForPoolMembers,
+        _counterForBondedPools,
+        _counterForRewardPools,
+        _maxPoolMembers,
+        _maxPoolMembersPerPool,
+        _maxPools,
+        _minCreateBond,
+        _minJoinBond,
+      ]: any) => {
+        // format optional configs to BN or null
+        _maxPoolMembers = _maxPoolMembers.toHuman();
+        if (_maxPoolMembers !== null) {
+          _maxPoolMembers = new BN(rmCommas(_maxPoolMembers));
+        }
+        _maxPoolMembersPerPool = _maxPoolMembersPerPool.toHuman();
+        if (_maxPoolMembersPerPool !== null) {
+          _maxPoolMembersPerPool = new BN(rmCommas(_maxPoolMembersPerPool));
+        }
+        _maxPools = _maxPools.toHuman();
+        if (_maxPools !== null) {
+          _maxPools = new BN(rmCommas(_maxPools));
+        }
+
+        setPoolsConfig({
+          ...poolsConfig,
+          stats: {
+            counterForPoolMembers: _counterForPoolMembers.toBn(),
+            counterForBondedPools: _counterForBondedPools.toBn(),
+            counterForRewardPools: _counterForRewardPools.toBn(),
+            maxPoolMembers: _maxPoolMembers,
+            maxPoolMembersPerPool: _maxPoolMembersPerPool,
+            maxPools: _maxPools,
+            minCreateBond: _minCreateBond.toBn(),
+            minJoinBond: _minJoinBond.toBn(),
+          },
+        });
+      }
+    );
+    setPoolsConfig({
+      ...poolsConfig,
+      unsub,
+    });
+  };
+
+  return (
+    <PoolsConfigContext.Provider
+      value={{
+        enabled,
+        stats: poolsConfig.stats,
+      }}
+    >
+      {children}
+    </PoolsConfigContext.Provider>
+  );
+};
