+++ src/modals/UpdateBond/Forms/BondSome.tsx
@@ -11,6 +11,7 @@ import { useSubmitExtrinsic } from 'library/Hooks/useSubmitExtrinsic';
 import { APIContextInterface } from 'types/api';
 import { ConnectContextInterface } from 'types/connect';
 import { usePools } from 'contexts/Pools';
+import { PoolsContextState } from 'types/pools';
 import { NotesWrapper } from '../../Wrappers';
 import { FormFooter } from './FormFooter';
 
@@ -22,7 +23,7 @@ export const BondSome = (props: any) => {
   const { setStatus: setModalStatus, setResize, config }: any = useModal();
   const { activeAccount } = useConnect() as ConnectContextInterface;
   const { getBondOptions }: any = useBalances();
-  const { getPoolBondOptions } = usePools();
+  const { getPoolBondOptions } = usePools() as PoolsContextState;
   const { target } = config;
 
   const stakeBondOptions = getBondOptions(activeAccount);
