+++ src/pages/Pools/Roles.tsx
@@ -10,6 +10,7 @@ import { CardHeaderWrapper } from 'library/Graphs/Wrappers';
 import { OpenAssistantIcon } from 'library/OpenAssistantIcon';
 import { usePools } from 'contexts/Pools';
 import { ConnectContextInterface } from 'types/connect';
+import { PoolsContextState } from 'types/pools';
 import { PoolAccount } from './PoolAccount';
 import { RolesWrapper } from './ManagePool/Wrappers';
 
@@ -17,7 +18,7 @@ export const Roles = () => {
   const { isReady } = useApi() as APIContextInterface;
   const { activeAccount } = useConnect() as ConnectContextInterface;
   const { fetchAccountMetaBatch } = useAccount();
-  const { activeBondedPool } = usePools();
+  const { activeBondedPool } = usePools() as PoolsContextState;
   const activePool = activeBondedPool;
   const { roles } = activePool || {};
