+++ src/pages/Pools/index.tsx
@@ -17,6 +17,7 @@ import { useApi } from 'contexts/Api';
 import { PoolList } from 'library/PoolList';
 import { usePools } from 'contexts/Pools';
 import { APIContextInterface } from 'types/api';
+import { PoolsContextState } from 'types/pools';
 import ActivePoolsStatBox from './Stats/ActivePools';
 import MinJoinBondStatBox from './Stats/MinJoinBond';
 import MinCreateBondStatBox from './Stats/MinCreateBond';
@@ -31,7 +32,7 @@ export const Pools = (props: PageProps) => {
   const { title } = page;
   const { network } = useApi() as APIContextInterface;
   const navigate = useNavigate();
-  const { bondedPools, isBonding, isNominator, targets } = usePools();
+  const { bondedPools, isBonding } = usePools() as PoolsContextState;
   const { isSyncing } = useUi();
 
   // back to overview if pools are not supported on network
